added isolate option to nexi

+++ Pytrithon/nexus.py
@@ -16,7 +16,8 @@ class After(Thread):
     self.after()
 
 class Nexus:
-  def __init__(self, name, host, port, master, config, after):
+  def __init__(self, name, host, port, master, isolate, config, after):
+    self.isolate = isolate
     self.config = config
     self.name = "#" if name is None else name
     self.nametree = Tree()
@@ -122,19 +123,21 @@ class Nexus:
     subprocess.Popen(["python", "agent", "-P", str(self.server.port)] + (["-d", str(delay)] if delay is not None else []) + (["-p", str(poll)] if poll is not None else []) + (["-e"] if edit else []) + (["-H"] if halt else []) + (["-s"] if secret else []) + (["-M"] if mute else []) + (["-E"] if not errors else []) + [agent] + args)
 
   def push_agent(self, agent, structure):  
-    filename = "workbench/fragments/"+agent[1:].replace(".", "/")+".ptf" if agent.startswith("$") else "workbench/agents/"+agent.replace(".", "/")+".pta"
-    os.makedirs("/".join(filename.split("/")[:-1]), exist_ok=True)
-    with open(filename, "w", encoding="utf-8") as f:
-      f.write(structure)
+    if not self.isolate:
+      filename = "workbench/fragments/"+agent[1:].replace(".", "/")+".ptf" if agent.startswith("$") else "workbench/agents/"+agent.replace(".", "/")+".pta"
+      os.makedirs("/".join(filename.split("/")[:-1]), exist_ok=True)
+      with open(filename, "w", encoding="utf-8") as f:
+        f.write(structure)
 
   def push_file(self, file, data):  
-    if os.path.abspath("workbench/"+file).replace("\\", "/").startswith(os.path.abspath("workbench").replace("\\", "/")):
-      os.makedirs(os.path.dirname("workbench/"+file), exist_ok=True)
-      if not data.strip():
-        try:
-          os.remove("workbench/"+file)
-        except FileNotFoundError:
-          pass
-      else:
-        with open("workbench/"+file, "wb") as f:
-          f.write(data)
+    if not self.isolate:
+      if os.path.abspath("workbench/"+file).replace("\\", "/").startswith(os.path.abspath("workbench").replace("\\", "/")):
+        os.makedirs(os.path.dirname("workbench/"+file), exist_ok=True)
+        if not data.strip():
+          try:
+            os.remove("workbench/"+file)
+          except FileNotFoundError:
+            pass
+        else:
+          with open("workbench/"+file, "wb") as f:
+            f.write(data)
